chore: move variable types
