+++ ui/src/timeMachine/components/variableToolbar/VariableTooltipContents.tsx
@@ -25,9 +25,8 @@ import {
 import {toComponentStatus} from 'src/shared/utils/toComponentStatus'
 
 // Types
-import {RemoteDataState} from 'src/types'
+import {RemoteDataState, VariableValues} from 'src/types'
 import {AppState} from 'src/types'
-import {VariableValues} from 'src/variables/types'
 
 interface StateProps {
   values?: VariableValues
