+++ ui/src/types/variables.ts
@@ -5,3 +5,28 @@ import {Label} from 'src/types'
 export interface Variable extends GenVariable {
   labels: Label[]
 }
+
+export type FluxColumnType =
+  | 'boolean'
+  | 'unsignedLong'
+  | 'long'
+  | 'double'
+  | 'string'
+  | 'base64Binary'
+  | 'dateTime'
+  | 'duration'
+
+export interface VariableValues {
+  values: string[]
+  valueType: FluxColumnType
+  selectedValue: string
+  error?: string
+}
+
+export interface VariableValuesByID {
+  [variableID: string]: VariableValues
+}
+
+export interface ValueSelections {
+  [variableID: string]: string
+}
