+++ ui/src/variables/selectors/index.tsx
@@ -18,12 +18,10 @@ import {
   VariableArguments,
   VariableArgumentType,
   Variable,
-} from 'src/types'
-import {
   VariableValues,
   VariableValuesByID,
   ValueSelections,
-} from 'src/variables/types'
+} from 'src/types'
 
 type VariablesState = AppState['variables']['variables']
 type ValuesState = AppState['variables']['values']['contextID']
