+++ ui/src/variables/utils/getVarAssignment.ts
@@ -2,8 +2,7 @@
 import {parseDuration} from 'src/shared/utils/duration'
 
 // Types
-import {VariableValues} from 'src/variables/types'
-import {VariableAssignment} from 'src/types/ast'
+import {VariableValues, VariableAssignment} from 'src/types'
 
 export const getVarAssignment = (
   name: string,
